pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["tensortours"]

[project]
name = "tensortours"
version = "0.1.0"
description = "Audio tour generation backend"
requires-python = ">=3.12"
authors = [
    {name = "TensorTours Team"}
]

dependencies = [
    "boto3>=1.28.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.3",
    "openai>=1.3.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.2",
    "googlemaps>=4.10.0",
    "psycopg2-binary>=2.9.0",
    "google-genai>=1.16.0",
    "google-adk>=1.0.0",
    "boto3-stubs[polly,s3,secretsmanager,sqs,lambda,dynamodb]>=1.28.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "moto>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest-flake8>=1.1.1",
    "pytest-black>=0.3.0",
    "pydantic-to-typescript>=2.0.0",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.autoflake]
remove_all_unused_imports = true
remove_unused_variables = true
remove_duplicate_keys = true
expand_star_imports = true

[tool.pytest.ini_options]
addopts = ""
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
filterwarnings = [
    "ignore::DeprecationWarning:botocore.*",
]
markers = [
    "flake8: Run flake8 on Python source files",
    "black: Run black on Python source files",
    "lint: Run all linting checks",
]

[tool.flake8]
max_line_length = 80
exclude = [".git", "__pycache__", "build", "dist", ".venv", "venv"]
ignore = ["E203", "W503", "E501", "W391"]  # Recommended when using Black
per_file_ignores = [
    "__init__.py:F401,F403",
    "tests/*:F401",
    "scripts/*:F401",
    "integration_tests/*:F401,F841"
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "pydantic.*"
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "boto3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "botocore.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mypy_boto3_dynamodb.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tensortours.lambda_handlers.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
warn_return_any = false
warn_unreachable = false
strict_optional = false
